update: fixed circular dependency issue, added spring and jwt dependencies and updated user entity with password field
