+++ src/main/java/com/big_joe/ojemba_bank/config/SecurityConfig.java
@@ -1,0 +1,4 @@
+package com.big_joe.ojemba_bank.config;
+
+public class SecurityConfig {
+}
