+++ src/main/java/com/big_joe/ojemba_bank/service/UserService.java
@@ -1,8 +1,12 @@
 package com.big_joe.ojemba_bank.service;
 
+import com.big_joe.ojemba_bank.data.model.Transactions;
 import com.big_joe.ojemba_bank.data.model.User;
 import com.big_joe.ojemba_bank.dto.*;
+import com.itextpdf.text.DocumentException;
 
+import java.io.FileNotFoundException;
+import java.util.List;
 import java.util.Optional;
 
 public interface UserService {
@@ -23,4 +27,5 @@ public interface UserService {
     BankResponse fundsTransfer(TransferRequest transferRequest);
 
     Optional<User> findByAccountNumber(String senderAccountNumber);
+    List<Transactions> generateAccountStatement(TransactionEnquiryReq req) throws DocumentException, FileNotFoundException;
 }
